+++ packages/soda-hooks/src/utils/compareArray.ts
@@ -4,6 +4,6 @@
  * @param b 数组b
  * @returns 是否相等
  */
-export function compareArray(a: any[], b: any[]) {
-    return a.length === b.length && a.every((value, index) => Object.is(value, b[index]))
+export function compareArray(a: any[] | readonly any[], b: any[] | readonly any[]) {
+    return Object.is(a, b) || (a.length === b.length && a.every((value, index) => Object.is(value, b[index])))
 }