+++ packages/soda-react-router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react-router
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.7.2
+
 ## 1.1.1
 
 ### Patch Changes